+++ templater/debug.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * Daniel Dias Rodrigues (aka Nerun) <danieldiasr@gmail.com>
+ */
+function printr($arr) {
+	echo "<pre>";
+	print_r($arr);
+	echo "</pre><br><br>";
+}
+?>